Make host url configurable - PMT #104364
The mediathread instance that the extension points to can now
be chosen with a dropdown menu in the extension's options.

For now, to make the UI simple, I'm just providing three pre-determined
choices for the mediathread url: production, staging, and qa.

I've done some refactoring here, removing settings.js since
it's not really necessary.

There's a bunch of changes to the common code that I will merge
into the rest of the repositories.