Make host_url completely configurable - PMT #105269
This changes the URL options to a radio button interface with an "Other"
field to enter a custom URL.

I've added an alert message that pops up if the is_logged_in ajax request
fails. This will let users know why the URL fails when entering a custom
url, and will help us debug any problems.

I've also incremented the version to 0.9.0 so I can make a new release.