Fix duplicate slash error when loading auth state from Mediathread
This error seems to have only started happening lately, otherwise we
would have ran into it sooner.

This may address #118, we'll see. When the extension is installed, the
hostUrl is set to `https://mediathread.ccnmtl.columbia.edu/`, with a
trailing slash, causing the request to break when the extension is used,
until the user sets the URL to the one without the trailing slash. This
change will fix that.